The great rewrite of 2018
A bunch of problems started to happen due to the order some things were loaded. Some plugins were not used as often while others outright broke for no good reason.

Most notably changes include the removal of several plugins that were not used, the change to fzf (from CtrlP) and init.vim is just a single file now (the plugins.d structure might have looked nicer but it was kind of a pain to maintain)